% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain-ladder.R
\name{create_chl_coefs_weighted_future}
\alias{create_chl_coefs_weighted_future}
\title{Volume-weighted CHL factors with weight at the \emph{current} column}
\usage{
create_chl_coefs_weighted_future(cum_trg, trg_weight, chl_length = "full")
}
\arguments{
\item{cum_trg}{Cumulative triangle.}

\item{trg_weight}{Weight matrix of the same shape as \code{cum_trg}.}

\item{chl_length}{As in \code{\link[=create_chl_coefs]{create_chl_coefs()}}.}
}
\value{
A \code{data.frame} analogous to the output of \code{\link[=create_chl_coefs]{create_chl_coefs()}}.
}
\description{
Variant of \code{\link[=create_chl_coefs_weighted]{create_chl_coefs_weighted()}} that uses \code{trg_weight[, j]}
instead of \code{trg_weight[, j - 1]}, i.e. the weight evaluated at the
developed (forward-looking) column.
}
