% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles.R
\name{create_decumulative_triangle}
\alias{create_decumulative_triangle}
\title{De-cumulate a cumulative development triangle}
\usage{
create_decumulative_triangle(trg)
}
\arguments{
\item{trg}{A cumulative triangle.}
}
\value{
The incremental triangle.
}
\description{
Inverse of \code{\link[=create_cumulative_triangle]{create_cumulative_triangle()}}.
}
\examples{
cum_trg <- matrix(c(10, 30, 32,  20, 27, NA,  15, NA, NA), nrow = 3, byrow = TRUE)
create_decumulative_triangle(cum_trg)
}
