% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_avg_coefs}
\alias{create_avg_coefs}
\title{Průměrné vývojové faktory napříč diagonálami}
\usage{
create_avg_coefs(chl_trg, avg_length = "full")
}
\arguments{
\item{chl_trg}{\code{matrix} vývojových faktorů (např. z \code{\link{create_chl_coef_triangle}}).}

\item{avg_length}{Vektor \code{character}/\code{numeric}; délky oken, \code{"full"} = všechny dostupné.}
}
\value{
\code{data.frame} s řádky = nastavení délek a sloupci = vývojové věky (0,1,2,\dots).
}
\description{
Z diagonální reprezentace trojúhelníku vývojových faktorů \code{chl_trg} počítá průměry
po řádcích přes \code{avg_length} posledních diagonál.
}
\seealso{
\code{\link{create_chl_coef_triangle}}, \code{\link{diag_reader}}
}
