% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_chl_coef_triangle}
\alias{create_chl_coef_triangle}
\title{Trojúhelník vývojových faktorů (link ratios) z kumulativního trojúhelníku}
\usage{
create_chl_coef_triangle(cum_trg)
}
\arguments{
\item{cum_trg}{\code{matrix} kumulativního trojúhelníku.}
}
\value{
\code{matrix} se stejnými rozměry jako \code{cum_trg} s vývojovými faktory.
}
\description{
Trojúhelník vývojových faktorů (link ratios) z kumulativního trojúhelníku
}
\details{
Pro \eqn{j>1} a existující hodnoty platí \eqn{DF_{i,j} = C_{i,j}/C_{i,j-1}}.
První sloupec je \code{NA}. Buňky za „okrajem“ trojúhelníku jsou \code{NA}.
}
