% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_chl_coefs}
\alias{create_chl_coefs}
\title{Výpočet průměrných CH-L koeficientů po sloupcích}
\usage{
create_chl_coefs(cum_trg, chl_length = "full")
}
\arguments{
\item{cum_trg}{\code{matrix} kumulativního trojúhelníku.}

\item{chl_length}{Vektor \code{character}/\code{numeric}; délky oken (např. \code{c("full", 3, 5)}).}
}
\value{
\code{data.frame} s řádky = nastavení délek a sloupci = vývojové věky (0,1,2,\dots).
První sloupec (věk 0) je \code{NA}.
}
\description{
Počítá pro každý sloupec (vývojový věk) poměr součtu kumulativ za \code{j} a \code{j-1}
v posledních \code{act_length} diagonálách. \code{"full"} znamená použít všechny dostupné diagonály.
}
