% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_cumulative_triangle}
\alias{create_cumulative_triangle}
\title{Vytvoří kumulativní trojúhelník}
\usage{
create_cumulative_triangle(trg)
}
\arguments{
\item{trg}{\code{matrix}. Trojúhelník v nekumulativní podobě.}
}
\value{
\code{matrix}. Trojúhelník v kumulativní podobě.
}
\description{
Funkce převede trojúhelník z nekumulativní podoby na kumulativní,
tj. v každém řádku sčítá hodnoty po směru sloupců.
}
