% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_decumulative_triangle}
\alias{create_decumulative_triangle}
\title{Vytvoří dekumulovaný trojúhelník}
\usage{
create_decumulative_triangle(trg)
}
\arguments{
\item{trg}{\code{matrix}. Trojúhelník v kumulativní podobě.}
}
\value{
\code{matrix}. Trojúhelník v nekumulativní podobě.
}
\description{
Funkce převede trojúhelník z kumulativní podoby zpět na nekumulativní,
tj. odečítá předchozí sloupce.
}
