% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_run_off_check}
\alias{create_run_off_check}
\title{Run-off kontrola kumulativního trojúhelníku}
\usage{
create_run_off_check(cum_trg)
}
\arguments{
\item{cum_trg}{\code{matrix} kumulativního trojúhelníku.}
}
\value{
Numerický vektor rozdílů (od nejstarší po nejnovější).
}
\description{
Porovnává součty hlavní diagonály s předchozími diagonálami a vrací rozdíly
(indikativní míra nedosběru/rozběhu).
}
\details{
Pro každou diagonálu kromě hlavní se počítá rozdíl součtu „ocasu“ hlavní diagonály
a součtu dané diagonály.
}
\seealso{
\code{\link{diag_reader}}
}
