% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{create_triangle}
\alias{create_triangle}
\title{Vytvoření trojúhelníku z primárních dat}
\usage{
create_triangle(
  data,
  row_num,
  col_num,
  value,
  cond_variable = NULL,
  cond_value = NULL,
  rows = NULL
)
}
\arguments{
\item{data}{\code{data.frame} – zdrojová data.}

\item{row_num}{\code{character} – jméno proměnné určující čísla řádků (obvykle vznik roku).}

\item{col_num}{\code{character} – jméno proměnné určující čísla sloupců (obvykle rok vývoje).}

\item{value}{\code{character} – jméno proměnné s hodnotami, které se sčítají.}

\item{cond_variable}{\code{character} (volitelně) – jméno jedné nebo více kategoriálních proměnných,
podle kterých se filtrují data.}

\item{cond_value}{\code{character} (volitelně) – hodnoty proměnných zadaných v \code{cond_variable}.}

\item{rows}{\code{integer} (volitelně) – maximální počet řádků matice. Pokud není zadán,
použije se maximum z \code{row_num}.}
}
\value{
Matice (\code{matrix}) o rozměru \code{rows x rows} s trojúhelníkovou strukturou.
}
\description{
Funkce vezme zdrojová data a podle zadaných parametrů vytvoří trojúhelníkovou matici
(běžně používanou v aktuárských aplikacích).
}
\examples{
\dontrun{
library(dplyr)
df <- data.frame(
  accident = c(1,1,2,2,3,3),
  dev = c(0,1,0,1,0,1),
  value = c(100, 50, 200, 80, 150, 60)
)

create_triangle(df,
                row_num = "accident",
                col_num = "dev",
                value   = "value")
}
}
