% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{diag_reader}
\alias{diag_reader}
\title{Načte diagonálu z trojúhelníku}
\usage{
diag_reader(trg, diag_num = nrow(trg))
}
\arguments{
\item{trg}{\code{matrix}. Trojúhelník.}

\item{diag_num}{\code{numeric}. Číslo diagonály (počítáno shora). Výchozí je počet řádků trojúhelníku (hlavní diagonála).}
}
\value{
\code{numeric}. Vektor hodnot na zvolené diagonále.
}
\description{
Funkce vybere hodnoty z dané diagonály trojúhelníku a převede je na vektor.
}
