% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{diag_writer}
\alias{diag_writer}
\title{Dopsání poslední diagonály z nového trojúhelníku}
\usage{
diag_writer(old_trg, new_trg)
}
\arguments{
\item{old_trg}{matrix/data.frame se starým trojúhelníkem.}

\item{new_trg}{matrix/data.frame s novým trojúhelníkem (o 1 delší hrana).}
}
\value{
\code{matrix} rozměru \code{new_trg} s doplněnou poslední diagonálou.
}
\description{
Rozšíří starý trojúhelník o řádek a sloupec s \code{NA} a doplní je
hodnotami z nového trojúhelníku po diagonále (aktuální období).
}
