% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangles_functions.R
\name{triangle_namer}
\alias{triangle_namer}
\title{Pojmenování řádků a sloupců trojúhelníku}
\usage{
triangle_namer(trg, claim_period_names)
}
\arguments{
\item{trg}{\code{matrix} trojúhelníku.}

\item{claim_period_names}{\code{character} vektor názvů (např. roky/čtvrtletí).}
}
\value{
\code{matrix} s nastavenými \code{rownames} a \code{colnames}.
}
\description{
Nastaví názvy řádků dle \code{claim_period_names} a názvy sloupců jako \code{0:(ncol(trg)-1)}.
}
